fix deadzone value casting to int instead of double

+++ PS4Macro/Forms/RemapperForm.cs
@@ -391,7 +391,7 @@ namespace PS4Macro.Forms
 
         private void decayThresholdNumericUpDown_ValueChanged(object sender, EventArgs e)
         {
-            Remapper.MouseDecayThreshold = (int)decayThresholdNumericUpDown.Value;
+            Remapper.MouseDecayThreshold = (double)decayThresholdNumericUpDown.Value;
         }
 
         private void deadzoneNumericUpDown_ValueChanged(object sender, EventArgs e)
